--- src/components/Admin.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-
-export default class Admin extends Component {
-  render() {
-    return (
-      <div className='row'>
-        <div className='col-md-12'>Раздел /admin</div>
-      </div>
-    )
-  }
-}
