+++ src/components/Genre.js
@@ -1,11 +1,10 @@
 import React, { Component } from 'react'
+import  '../style/stats.scss';
 
-export default class Genre extends Component {
+export default class Stats extends Component {
   render() {
     return (
-      <div className='row'>
-        <div className='col-md-12'>Раздел /genre</div>
-      </div>
+      
     )
   }
 }
